requirements.txt
python-dotenv==1.0.0
langchain==0.0.352
langchain-openai==0.0.5
tiktoken
chromadb
